make relation between models and start to make accunting service

+++ database/migrations/2024_11_02_210344_create_products_table.php
@@ -19,8 +19,10 @@ return new class extends Migration
             $table->text("body");
             $table->json("jsonld");
 
-            $table->unsignedSmallInteger("category_id");
-            $table->foreign("category_id")->references("id")->on("categories");
+            $table->unsignedBigInteger("category_id");
+            $table->foreign("category_id")
+                ->references("id")
+                ->on("categories");
 
             $table->timestamps();
         });
